requirements.txt
#############################################
# Project dependencies for smarthr360_backend
# Generated from the codebase (Django + DRF + JWT)
#############################################

# Core web framework
# Pinned to the Django version used to generate the project
Django==5.2.8

# Django REST framework (used across views/serializers/tests)
djangorestframework>=3.15,<4.0

# JWT authentication for DRF (used in settings and views)
djangorestframework-simplejwt>=5.3,<6.0

# API documentation with OpenAPI/Swagger
drf-spectacular>=0.27,<1.0
